Moved RoundResultEnum to model package
