+++ src/main/java/com/hartwig/paperrockscissors/model/HumanPlayer.java
@@ -2,6 +2,10 @@ package com.hartwig.paperrockscissors.model;
 
 import com.hartwig.paperrockscissors.engine.GameControls;
 
+/**
+* Human Player using GameControls to identify move
+*
+*/
 public class HumanPlayer implements PlayerInterface {
   private final GameControls gameControls;
