+++ src/main/java/com/hartwig/paperrockscissors/model/RoundResultEnum.java
@@ -1,4 +1,4 @@
-package com.hartwig.paperrockscissors.engine;
+package com.hartwig.paperrockscissors.model;
 
 public enum RoundResultEnum {
   PLAYER1_WINS,
